Conflicts:
	src/core/lib/core_common/signal.hpp
	src/core/plugins/plg_qt_common/CMakeLists.txt

+++ src/core/testing/plg_ui_main_test/pages/test_page.cpp
@@ -138,7 +138,7 @@ void TestPage::getTextField( std::wstring * text ) const
 	*text = text_;
 }
 
-void TestPage::setSlideData( const int & length )
+void TestPage::setSlideData( const double & length )
 {
 	if ((length < this->getSlideMinData()) || (length > this->getSlideMaxData()))
 	{
@@ -146,7 +146,7 @@ void TestPage::setSlideData( const int & length )
 	}
 	curSlideData_ = length;
 }
-void TestPage::getSlideData( int * length ) const
+void TestPage::getSlideData(double * length) const
 {
 	*length = curSlideData_;
 }
